+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #### Features ####
 *   Add support for key modifiers for send_keys [Issue #420] (Sarah Mogin)
+*   Drag by offset support in native element (phoenixek12)
 
 #### Bug fixes ####